Permitir bloque/tema_legacy como texto (no int)
En el SQLite original la columna bloque se declaró INTEGER pero
SQLite es de tipos dinámicos y hay registros con texto como
"Punto de Acceso General (PAG)".  Postgres rechaza estos valores
con int estricto.  Cambiados bloque y tema_legacy a text en el
schema, en la función importar_test y en el migrador.

Los temas "reales" siguen viviendo en la tabla 'temas' con
embeddings; bloque/tema_legacy son solo etiquetas heredadas.

+++ db/init/03_funciones.sql
@@ -159,8 +159,8 @@ BEGIN
         VALUES (
             v_preg->>'pregunta',
             v_preg->'opciones',
-            NULLIF(v_preg->>'bloque','')::int,
-            NULLIF(v_preg->>'tema','')::int,
+            NULLIF(v_preg->>'bloque',''),
+            NULLIF(v_preg->>'tema',''),
             jwt_usuario_id()
         )
         ON CONFLICT (hash_contenido) DO UPDATE
